@@ -1,7 +1,11 @@
 """
 memory/scripts/memory.py - Memory Skill Commands
 
-Phase 63: Migrated from tools.py to scripts pattern.
+Modernized:
+- @skill_command with autowire=True for clean dependency injection
+- Uses PRJ_DATA_HOME/memory for persistent storage (not PRJ_CACHE)
+- Uses ConfigPaths for semantic path resolution
+- Refactored to use omni.foundation for embedding and vector_store.
 """
 
 import json
@@ -10,103 +14,60 @@ import uuid
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
-import structlog
+from omni.foundation.api.decorators import skill_command
+from omni.foundation.config.paths import ConfigPaths
+from omni.foundation.config.logging import get_logger
+from omni.foundation.services.embedding import get_embedding_service
+from omni.foundation.services.vector import get_vector_store
 
-from agent.skills.decorators import skill_command
-
-logger = structlog.get_logger(__name__)
+logger = get_logger("skill.memory")
 
 # =============================================================================
-# Vector Store Initialization (Rust + LanceDB)
+# Vector Store Initialization (Foundation Layer)
 # =============================================================================
 
 
 def _get_memory_path() -> Path:
-    """Get memory root path."""
-    from common import prj_dirs
-    from common.settings import get_setting
-
-    # Check for custom override in settings.yaml
-    custom_path = get_setting("memory.path", "")
-    if custom_path:
-        return Path(custom_path)
-
-    return prj_dirs.get_cache_dir("memory")
+    """Get memory root path using PRJ_DATA_HOME."""
+    # Use ConfigPaths for semantic path resolution (Layer 1)
+    paths = ConfigPaths()
+    # Memory data goes in $PRJ_DATA_HOME/memory (persistent data, not cache)
+    return paths.get_data_dir("memory")
 
 
 MEMORY_ROOT = _get_memory_path()
-DB_PATH = MEMORY_ROOT / "lancedb"
-DB_PATH.mkdir(parents=True, exist_ok=True)
-
-# Lazy initialization for VectorStore (avoid re-initialization on repeated imports)
-_cached_store: Optional[Any] = None
-RUST_AVAILABLE = False
-
-
-def _get_store() -> Optional[Any]:
-    """Lazily initialize and return the VectorStore."""
-    global _cached_store, RUST_AVAILABLE
-
-    if _cached_store is not None:
-        return _cached_store
-
-    try:
-        import omni_core_rs
-
-        # Get embedding dimension from the embedding service
-        # This ensures Rust LanceDB schema matches Python embedding dimension
-        embedding_dimension = 384  # Default for BGE-small
-        try:
-            from agent.core.embedding import get_embedding_service
-
-            embedding_dimension = get_embedding_service().dimension
-        except Exception:
-            pass  # Use default dimension
-
-        _cached_store = omni_core_rs.PyVectorStore(str(DB_PATH), embedding_dimension)
-        RUST_AVAILABLE = True
-        # Only log on first initialization
-        logger.debug(
-            f"Librarian (VectorStore) ready at {DB_PATH} with dimension {embedding_dimension}"
-        )
-    except Exception as e:
-        RUST_AVAILABLE = False
-        _cached_store = None
-        logger.warning(f"Failed to init Rust VectorStore: {e}")
-
-    return _cached_store
-
-
-# Default table name
 DEFAULT_TABLE = "knowledge_base"
 
 
 def _get_embedding(text: str) -> List[float]:
     """
-    [Phase 53.5] Real Semantic Embedding using FastEmbed or OpenAI.
+    Get embedding using Foundation embedding service.
 
-    Falls back to DummyEmbedding if real embedding service fails.
+    Falls back to deterministic dummy embedding if service fails.
     """
     try:
-        from agent.core.embedding import get_embedding_service
-
         service = get_embedding_service()
         return service.embed(text)[0]
     except Exception as e:
-        logger.warning(f"Real embedding failed, using dummy: {e}")
-        # Fallback to dummy embedding for testing
-        seed = sum(ord(c) for c in text) % 100
-        dimension = 384  # BGE-small default dimension
-        return [0.001 * seed] * dimension
+        logger.warning(f"Embedding service failed, using fallback: {e}")
+        # Fallback to deterministic hash-based embedding
+        import hashlib
+
+        hash_bytes = hashlib.sha256(text.encode()).digest()
+        vector = [float(b) / 255.0 for b in hash_bytes]
+        dimension = 1536  # Default fallback dimension
+        repeats = (dimension + len(vector) - 1) // len(vector)
+        return (vector * repeats)[:dimension]
 
 
 def _load_skill_manifest(skill_name: str) -> tuple[Optional[Dict[str, Any]], Optional[str]]:
     """Load a skill's manifest and prompts."""
-    from common.settings import get_setting
-    from common.gitops import get_project_root
+    from omni.foundation.config import get_setting
+
+    paths = ConfigPaths()
+    project_root = paths.project_root
 
     skills_path = get_setting("skills.path", "assets/skills")
-    project_root = get_project_root()
     skills_dir = project_root / skills_path
     skill_path = skills_dir / skill_name
 
@@ -147,9 +108,13 @@ def _load_skill_manifest(skill_name: str) -> tuple[Optional[Dict[str, Any]], Opt
     name="save_memory",
     category="write",
     description="Store a key insight into long-term memory.",
-    inject_settings=["memory.path"],
+    autowire=True,
 )
-async def save_memory(content: str, metadata: Optional[Dict[str, Any]] = None) -> str:
+async def save_memory(
+    content: str,
+    metadata: Optional[Dict[str, Any]] = None,
+    paths: ConfigPaths | None = None,
+) -> str:
     """
     [Long-term Memory] Store a key insight, decision, or learning into LanceDB.
 
@@ -165,13 +130,13 @@ async def save_memory(content: str, metadata: Optional[Dict[str, Any]] = None) -
     Returns:
         Confirmation message with stored content preview
     """
-    store = _get_store()
-    if not RUST_AVAILABLE or not store:
-        return "Rust VectorStore not available. Cannot store memory."
+    client = get_vector_store()
+    store = client.store
+    if not store:
+        return "VectorStore not available. Cannot store memory."
 
     try:
         doc_id = str(uuid.uuid4())
-        vector = _get_embedding(content)
 
         # [FIX] Robust metadata handling - handle str, None, or dict
         if metadata is None:
@@ -189,8 +154,10 @@ async def save_memory(content: str, metadata: Optional[Dict[str, Any]] = None) -
         # Add timestamp to metadata (after ensuring it's a dict)
         metadata["timestamp"] = time.time()
 
-        store.add_documents(DEFAULT_TABLE, [doc_id], [vector], [content], [json.dumps(metadata)])
-        return f"Saved memory [{doc_id[:8]}]: {content[:80]}..."
+        success = await client.add(content, metadata, collection=DEFAULT_TABLE)
+        if success:
+            return f"Saved memory [{doc_id[:8]}]: {content[:80]}..."
+        return "Failed to store memory."
     except Exception as e:
         logger.error("save_memory failed", error=str(e))
         return f"Error saving memory: {str(e)}"
@@ -200,8 +167,13 @@ async def save_memory(content: str, metadata: Optional[Dict[str, Any]] = None) -
     name="search_memory",
     category="read",
     description="Semantically search memory for relevant past experiences.",
+    autowire=True,
 )
-async def search_memory(query: str, limit: int = 5) -> str:
+async def search_memory(
+    query: str,
+    limit: int = 5,
+    paths: ConfigPaths | None = None,
+) -> str:
     """
     [Retrieval] Semantically search memory for relevant past experiences or rules.
 
@@ -217,30 +189,18 @@ async def search_memory(query: str, limit: int = 5) -> str:
     Returns:
         Relevant memories found, or "No relevant memories found"
     """
-    store = _get_store()
-    if not RUST_AVAILABLE or not store:
-        return "Rust VectorStore not available. Cannot search memory."
-
     try:
-        vector = _get_embedding(query)
+        results = await get_vector_store().search(query, n_results=limit, collection=DEFAULT_TABLE)
 
-        # Call Rust search - returns JSON strings
-        results_json = store.search(DEFAULT_TABLE, vector, limit)
-
-        if not results_json:
+        if not results:
             return "No matching memories found."
 
-        output = [f"Found {len(results_json)} matches for '{query}':"]
-        for r_str in results_json:
-            r = json.loads(r_str)
-            dist = r.get("distance", 0.0)
-            content = r.get("content", "")
-            meta = r.get("metadata", {})
-
+        output = [f"Found {len(results)} matches for '{query}':"]
+        for r in results:
             # Format output for LLM consumption
-            output.append(f"- [Score: {dist:.4f}] {content[:100]}")
-            if meta:
-                output[-1] += f" (Meta: {json.dumps(meta)[:50]}...)"
+            output.append(f"- [Score: {r.distance:.4f}] {r.content[:100]}")
+            if r.metadata:
+                output[-1] += f" (Meta: {json.dumps(r.metadata)[:50]}...)"
 
         return "\n".join(output)
     except Exception as e:
@@ -252,8 +212,11 @@ async def search_memory(query: str, limit: int = 5) -> str:
     name="index_memory",
     category="write",
     description="Optimize memory index for faster search (IVF-FLAT).",
+    autowire=True,
 )
-async def index_memory() -> str:
+async def index_memory(
+    paths: ConfigPaths | None = None,
+) -> str:
     """
     [Optimization] Create/optimize vector index for faster search.
 
@@ -263,13 +226,11 @@ async def index_memory() -> str:
     Returns:
         Confirmation of index creation
     """
-    store = _get_store()
-    if not RUST_AVAILABLE or not store:
-        return "Rust VectorStore not available. Cannot create index."
-
     try:
-        store.create_index(DEFAULT_TABLE)
-        return "Index creation/optimization complete. Search performance improved."
+        success = await get_vector_store().create_index(collection=DEFAULT_TABLE)
+        if success:
+            return "Index creation/optimization complete. Search performance improved."
+        return "Failed to create index."
     except Exception as e:
         return f"Error creating index: {str(e)}"
 
@@ -278,20 +239,19 @@ async def index_memory() -> str:
     name="get_memory_stats",
     category="view",
     description="Get statistics about stored memories.",
+    autowire=True,
 )
-async def get_memory_stats() -> str:
+async def get_memory_stats(
+    paths: ConfigPaths | None = None,
+) -> str:
     """
     [Diagnostics] Get statistics about stored memories.
 
     Returns:
         Count of stored memories
     """
-    store = _get_store()
-    if not RUST_AVAILABLE or not store:
-        return "Rust VectorStore not available."
-
     try:
-        count = store.count(DEFAULT_TABLE)
+        count = await get_vector_store().count(collection=DEFAULT_TABLE)
         return f"Stored memories: {count}"
     except Exception as e:
         return f"Error getting stats: {str(e)}"
@@ -301,8 +261,12 @@ async def get_memory_stats() -> str:
     name="load_skill",
     category="write",
     description="Load a skill's manifest into semantic memory.",
+    autowire=True,
 )
-async def load_skill(skill_name: str) -> str:
+async def load_skill(
+    skill_name: str,
+    paths: ConfigPaths | None = None,
+) -> str:
     """
     [Skill Loader] Load a single skill's manifest into semantic memory.
 
@@ -315,9 +279,10 @@ async def load_skill(skill_name: str) -> str:
     Returns:
         Confirmation message with skill details
     """
-    store = _get_store()
-    if not RUST_AVAILABLE or not store:
-        return "Rust VectorStore not available. Cannot load skill."
+    client = get_vector_store()
+    store = client.store
+    if not store:
+        return "VectorStore not available. Cannot load skill."
 
     manifest, prompts = _load_skill_manifest(skill_name)
     if not manifest:
@@ -343,29 +308,30 @@ async def load_skill(skill_name: str) -> str:
         document += f"\n---\n\n## System Prompts\n{prompts[:2000]}"
 
     doc_id = f"skill_{skill_name}"
-    vector = _get_embedding(document)
-
-    store = _get_store()
-    if not RUST_AVAILABLE or not store:
-        return "Rust VectorStore not available. Cannot load skill."
 
     try:
-        store.add_documents(
-            DEFAULT_TABLE,
-            [doc_id],
-            [vector],
-            [document],
-            [
-                json.dumps(
-                    {
-                        "type": "skill_manifest",
-                        "skill_name": skill_name,
-                        "version": manifest.get("version", "unknown"),
-                    }
-                )
-            ],
+        success = await client.add(
+            document,
+            metadata={
+                "type": "skill_manifest",
+                "skill_name": skill_name,
+                "version": manifest.get("version", "unknown"),
+            },
+            collection=DEFAULT_TABLE,
         )
-
-        return f"✅ Skill '{skill_name}' loaded into semantic memory."
+        if success:
+            return f"Skill '{skill_name}' loaded into semantic memory."
+        return f"Failed to load skill '{skill_name}'."
     except Exception as e:
         return f"Failed to load skill '{skill_name}': {e}"
+
+
+__all__ = [
+    "save_memory",
+    "search_memory",
+    "index_memory",
+    "get_memory_stats",
+    "load_skill",
+    "MEMORY_ROOT",
+    "DEFAULT_TABLE",
+]
