@@ -12,13 +12,14 @@ Contains atomic implementations for skill-related commands:
 
 from __future__ import annotations
 
-from agent.skills.skill.scripts.discovery import (
+# Import from local modules (v2.0 pattern)
+from skill.scripts.discovery import (
     discover,
     suggest,
     jit_install,
     list_index,
 )
-from agent.skills.skill.scripts.templates import (
+from skill.scripts.templates import (
     list_templates,
     get_template_info,
     get_template_source,
@@ -27,10 +28,10 @@ from agent.skills.skill.scripts.templates import (
     format_eject_result,
     format_info_result,
 )
-from agent.skills.skill.scripts.list_tools import list_tools, format_tools_list
-from agent.skills.skill.scripts.search_tools import search_tools, format_search_result
-from agent.skills.skill.scripts.unload import unload_skill
-from agent.skills.skill.scripts.reload import reload_skill
+from skill.scripts.list_tools import list_tools, format_tools_list
+from skill.scripts.search_tools import search_tools, format_search_result
+from skill.scripts.unload import unload_skill
+from skill.scripts.reload import reload_skill
 
 __all__ = [
     # Discovery commands
