@@ -1,6 +1,6 @@
 //! Batch Refactoring Engine - Heavy-Duty Parallel Processing
 //!
-//! Phase 58: The Ouroboros - Self-Eating Snake
+//! The Ouroboros - Self-Eating Snake
 //!
 //! Provides parallel batch refactoring across entire codebases using rayon
 //! and ignore for maximum performance. Python sends one command, Rust
