@@ -3,8 +3,8 @@
 //! omni-edit - Structural Code Refactoring for Omni DevEnv
 //!
 //! A high-performance AST-based code modification library using ast-grep.
-//! Part of Phase 52: The Surgeon (CCA-Aligned Code Modification).
-//! Phase 58: The Ouroboros - Heavy-duty parallel batch refactoring.
+//! Part of The Surgeon (CCA-Aligned Code Modification).
+//! The Ouroboros - Heavy-duty parallel batch refactoring.
 //!
 //! # Features
 //!
@@ -12,7 +12,7 @@
 //! - **Diff Preview**: Generate unified diffs before applying changes
 //! - **Multi-language**: Python, Rust, JavaScript, TypeScript support
 //! - **Capture Substitution**: Use `$NAME` and `$$$` patterns
-//! - **Batch Refactoring**: Parallel processing across entire codebases (Phase 58)
+//! - **Batch Refactoring**: Parallel processing across entire codebases (The Ouroboros)
 //!
 //! # Architecture (ODF-REP Compliant)
 //!
@@ -24,7 +24,7 @@
 //! ├── diff.rs     # Diff generation utilities
 //! ├── capture.rs  # Variable capture substitution
 //! ├── editor.rs   # StructuralEditor implementation
-//! └── batch.rs    # Batch refactoring with rayon (Phase 58)
+//! └── batch.rs    # Batch refactoring with rayon (The Ouroboros)
 //! ```
 //!
 //! # Example
@@ -40,7 +40,7 @@
 //!     "python"
 //! )?;
 //!
-//! // Batch refactoring across codebase (Phase 58)
+//! // Batch refactoring across codebase (The Ouroboros)
 //! let stats = StructuralEditor::batch_replace(
 //!     "/project",
 //!     "print($ARGS)",
@@ -62,7 +62,7 @@ mod capture;
 mod diff;
 mod editor;
 mod error;
-mod types; // Phase 58: Heavy-duty batch refactoring
+mod types; // Heavy-duty batch refactoring
 
 // ============================================================================
 // Public Re-exports
@@ -72,7 +72,7 @@ pub use editor::StructuralEditor;
 pub use error::EditError;
 pub use types::{EditConfig, EditLocation, EditResult};
 
-// Phase 58: Batch refactoring exports
+// Batch refactoring exports
 pub use batch::{BatchConfig, BatchRefactorStats};
 
 // Re-export diff utility for external use
