@@ -1,6 +1,7 @@
 //! Index operations for the vector store.
 //!
 //! Provides vector index creation for ANN search optimization.
+//! Uses adaptive IVF-FLAT index with optimal partition count based on dataset size.
 
 use lance::dataset::Dataset;
 use lance::index::vector::VectorIndexParams;
@@ -10,11 +11,20 @@ use lance_linalg::distance::DistanceType;
 
 use crate::{VECTOR_COLUMN, VectorStoreError};
 
+/// Minimum vectors before index is useful
+const MIN_VECTORS_FOR_INDEX: usize = 100;
+/// Vectors per partition (heuristic)
+const VECTORS_PER_PARTITION: usize = 256;
+/// Max partitions to avoid over-sharding
+const MAX_PARTITIONS: usize = 512;
+
 impl crate::VectorStore {
     /// Create a vector index for a table.
     ///
-    /// Creates an IVF-FLAT index for efficient ANN search on large datasets.
-    /// For datasets with < 10k vectors, flat search is usually sufficient.
+    /// Creates an adaptive IVF-FLAT index with optimal partition count.
+    /// Partition count = min(max(vectors / 256, 32), 512)
+    ///
+    /// For small datasets (< 1k vectors), flat search is typically faster.
     ///
     /// # Arguments
     ///
@@ -34,8 +44,26 @@ impl crate::VectorStore {
             .await
             .map_err(VectorStoreError::LanceDB)?;
 
+        // Get row count for adaptive partition sizing
+        let num_rows = dataset
+            .count_rows(None)
+            .await
+            .map_err(VectorStoreError::LanceDB)? as usize;
+
+        // Skip indexing for very small datasets
+        if num_rows < MIN_VECTORS_FOR_INDEX {
+            log::info!(
+                "Table '{table_name}' has {num_rows} vectors (min: {}), skipping index creation",
+                MIN_VECTORS_FOR_INDEX
+            );
+            return Ok(());
+        }
+
+        // Calculate adaptive partition count
+        let num_partitions = (num_rows / VECTORS_PER_PARTITION).clamp(32, MAX_PARTITIONS);
+
         // Create IVF-FLAT index with L2 distance
-        let params = VectorIndexParams::ivf_flat(128, DistanceType::L2);
+        let params = VectorIndexParams::ivf_flat(num_partitions, DistanceType::L2);
 
         dataset
             .create_index(
@@ -48,7 +76,11 @@ impl crate::VectorStore {
             .await
             .map_err(VectorStoreError::LanceDB)?;
 
-        log::info!("Created vector index for table '{table_name}'");
+        log::info!(
+            "Created IVF-FLAT index for table '{table_name}' ({num_rows} vectors, {num_partitions} partitions)",
+            num_rows = num_rows,
+            num_partitions = num_partitions
+        );
 
         Ok(())
     }
